Add tests that enumerate and verify rewrite rules
This feature requires some minor modifications to the regression script
to ignore the output when only an exit status is given. Additionally,
the regression script avoids running with `--check-synth-sol` when
`--sygus-rr` is used. Finally, we now exit with an exit status of 2 when
an unsound rewrite is detected. This is to distinguish from the exit
status of 1 that we get when we are reaching the limits imposed by
`--sygus-abort-size`.