Refactor unconstrained simplification pass
Before this refactoring, the UnconstrainedSimplifier was a class in
src/theory. The refactoring turns it into a PreprocessingPass because
that was the only use case for it. The refactoring also transfers the
ownership of the instance of that class from TheoryEngine to SmtEngine
because the TheoryEngine was just handing through requests from the
SmtEngine to the UnconstrainedSimplifier but was not using the
UnconstrainedSimplifier otherwise.

Testing: There are several tests in regress0/unconstrained that use this
pass. Running the regressions after the refactoring remains fast and the
regressions pass but disabling the pass makes them run very slow. This
seems to indicate that the pass is working as intended.