Avoid setting filename when set-info :name is used
Currently, `SmtEngine::setInfo()` changes `SmtEngine::d_filename` if it
is called on `:name`. The SMT-LIB 2.6 standard, however, says that
`:name` corresponds to the name of the solver. We are following the
standard by returning the name of CVC4 when `(get-info :name)` is called
(even after calling `(set-info :name ...)`), so changing `d_filename` is
the wrong thing to do and a noop w.r.t. `(get-option :name)`. This
commit removes the assignment that changes `d_filename`. Note that
`SmtEngine::setInfo()` is only called from the SMT2 parser.