Use uniform length limit for String constants
This commit changes the code in the strings solver to use the
String::maxSize() length limit everywhere instead of having multiple
constants defined in different places. This fixes an issue in our
nightlies where we were trying to rewrite an indexof operation with a
start index greater than the maximum value of a 32-bit integer. The
rewriter was checking for a start index greater than the maximum of a
64-bit integer before trying to convert the Rational into a 32-bit
number, which failed with CLN. This commit fixes the check.