Unify rewrites like (str.contains x y) --> (= x y)
This commit unifies rewrites that turn `(str.contains x y)` into a
conjunction of equalities using length-based reasoning. It introduces a
generic helper function for determining which `ys` from an inequality
`y1 + ...  + yn >= x` may be 0 for it to still hold. When we have
`(str.contains x y)` and we can show that `str.len(y) >= str.len(x)`
then we know that `(str.contains x y) --> (= x y)` and we can use the
helper function to determine subterms in `y` that are empty if `y` is in
`x`, which allows us to introduce equalities of the form `(= yi "")`.
This approach unifies multiple existing rewrites. Additional rewrites
will be subsumed once we merge the approximations for arithmetic
entailment.