Skip invalid objects in ContextObj::destroy()
Background
==========

Context-dependent objects store their older version in the pointer
`d_pContextObjRestore`. Additionally, each object is part of a
doubly-linked list of objects in the same scope (i.e. the same level).
When a scope is popped, all objects in the linked list look up their
previous version and insert themselves into the linked list at the level
of the previous version. If an object does not have an older version,
the `d_pContextObjRestore` pointer is `null` and restoring the older
version does not do anything.

When we destroy an object (which should be called from the destructor of
a context-dependent object), remove the object from all scopes by
repeatedly removing it from the current scope, restoring the older
version, removing it from the older scope and so on.

A side note: The doubly-linked lists used are somewhat unusual: Instead
of holding a pointer to the previous object, the context-dependent
objects store a pointer to the next pointer of the previous object.

Problem
=======

The mechanisms described above may be problematic under rare
circumstances. Specifically, this came up because Clang 6.0's ASAN
detected a stack-user-after-scope issue in the
`testTopScopeContextObj()` unit test in context_black.h. Note that Clang
5.0's ASAN was not able to detect this issue. In this unit test the
following happened:

1. We push our context.
2. We create two context-dependent objects: `x` at level 1 (by setting
`allocatedInCMM` to true when calling the `ContextObj` constructor) and
`y` at level 0. At this point, scope 0 has `y` in its linked list and
scope 1 has `x` in its linked list.
3. We call `makeCurrent()` on both objects. This does nothing for `x`
because it already is at level 1. For `y` it creates a backup of the
current information and inserts `y` into the scope 1. Scope 0 now has a
backup of `y` in its linked list and scope 1 has both `x` and `y` in its
linked list. Note that `x`'s `prev()` now points to `y`'s `next()`.
4. We do an additional push, another call to `makeCurrent` on both
objects, and a first `pop()`. After those operations, we have the same
state as after 3.
5. We pop our context again, which restores the original `y` in scope 0.
We now only have scope 0 with `y` in it. As described above, `x` is
untouched because there is no older version to restore.
6. We are at the end of the function and because `y` was created after
`x`, it is destroyed first (standard destruction order of stack objects
in C++). When destroying `x`, `x`'s `prev()` still points to `y`'s
`next()`, so we are doing an invalid write and ASAN complains.

Solution
========

The solution that this commit implements is fairly simple. We can detect
that `x`'s `prev()`/`next()` pointers are stale by checking whether the
objects level is greater than the context's level. If it is, we know
that this object is invalid and that we do not have any old versions to
remove. Note that this situation should be very rare because objects
with `allocatedInCMM` should usually be allocated in context memory,
where they get cleaned up automatically when the context is popped below
the level that they are associated with.