Use https for antlr3.org downloads
If wget is not available, curl is used to download files in contrib/get-script-header.sh, sourced in e.g. contrib/get-antlr-3.4. Curl doesn't automatically redirect when given a HTTP 301 response, and the www.antlr3.org server serves such a redirect to the HTTPS version of the site when reached via plain HTTP.
This commit changes the two www,antlr3.org URL's in contrib/get-antlr-3.4 to use https instead of http, which makes them work when curl is chosen in `webget`.