Improve skolem caching by normalizing skolem args
In certain cases, we can share skolems between similar reductions, e.g.
`(str.replace x y z)` and `(str.replace (str.++ x w) y z)` because the
first occurrence of `y` in `(str.++ x w)` has to be the first occurrence
of `y` in `x` (assuming that `y` appears in both, otherwise the value of
the skolems does not matter). This commit adds a helper function in the
skolem cache that does some of those simplifications.