RNG: Avoid modulo bias in random numbers
Computing `rand() % range` is problematic because it is biased if the
range does not evenly divide the largest possible random value [0].
This commit uses `std::uniform_int_distribution` to generate a random
value within a given range, avoiding the problem. Similarly, generating
floating-point values in an unbiased matter is tricky, and the commit
uses `std::uniform_real_distribution` to avoid those issues. Note that
`std::uniform_real_distribution` generates values in the range [a, b).
The commit uses the suggested solution in [1] to generate values in [a,
b].

[0] https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle#Modulo_bias
[1] https://en.cppreference.com/w/cpp/numeric/random/uniform_real_distribution#Notes