Language bindings: Enable catching of exceptions
Fixes #2810. SWIG relies on `throw` specifiers to determine which
exceptions a method can throw. The wrappers generated by SWIG catch
those C++ exceptions and turn them into exceptions for the target
language. However, we have removed `throw` specifiers because they have
been deprecated in C++11, so SWIG did not know about any of our
exceptions. This commit fixes the issue using the `%catches` directive,
declaring that all methods may throw a `CVC4::Exception` or a general
exception. Note: This means that users of the language bindings will
just receive a general `CVC4::Exception` instead of more specific
exceptions like `TypeException`s. Given that we are planning to have a
single exception type for the new CVC4 API, this seemed like a natural
choice.
Additionally, the commit (significantly) simplifies the mapping of C++
to Java exceptions and fixes an issue with Python exceptions not
inheriting from `BaseException`. Finally, the commit adds API examples
for Java and Python, which demonstrate catching exceptions, and adds
  Python examples as tests in our build system.