CVC: Fix printing of tuple and record selectors
Previously, tuple and record selectors were printed as postifx
operations, e.g. `a(r)` instead of `r.a`. This commit fixes the issue.
The commit also adds a regression test. Note that the test uses
`--no-dt-share-sel` because shared selectors make
`getSelectorIndexInternal()` return -1, which should be fixed in a
future commit.